requirements.txt
flask 
beautifulsoup4 
markdown 
faiss-cpu 
scikit-learn 
nltk